pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "numpy>=2.2.4",
    "pillow>=11.1.0",
    "python-dotenv>=1.1.0",
    "tensorflow>=2.14.0",
    "tensorflowjs>=4.22.0",
]
